vinyl: make lazy vy_mem copying of out statements
vy_mem stores its tuples in lsregion. Such tuples
can't be referenced or unreferenced because are
stored in monolitic memory area.

Earlier vy_mem was making duplicates of the tuples
that must be returned to the merge iterator to avoid
referencing or unreferencing lsregion tuples.

Now vy_mem returns lsregion tuples and the merge
iterator makes the duplicates if it is necessary.

Need for #1908